test: ApiEventsProvider intialState behaviour
This commit adds tests for the initialization logic for the
ApiEventsProvider.

Note:
Testing the ApiEventsProvider directly should be considered testing
implementation details because the End User cannot see the Provider.

If you want to test the Provider, you should test it by proxy via a
Component that uses the Provider instead.

Cases:
- throws an error when an invalid initalState is provided
- does not throw error when initialState is not provided
- allows an ImmutableMap<string, ApiEvent> as the initialState
- allows a Map<string, ApiEvent> as the initialState
- allows an Array<[string, ApiEvent]> as the initialState